[MP-008] Refactoring: moved job data and metadata into nested packages.
