+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/handlers/JobHandler.kt
@@ -1,6 +1,6 @@
 package io.scheduler.comparison.quartz.jobs.handlers
 
-import io.scheduler.comparison.quartz.jobs.state.JobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.JobMetadata
 
 // Todo: replace JobData/JobMetadata with JobState aggregate interface
 fun interface JobHandler<T, V : JobMetadata> {
