+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/handlers/pagination/PaginatedJobHandlerBase.kt
@@ -1,7 +1,7 @@
 package io.scheduler.comparison.quartz.jobs.handlers.pagination
 
 import io.github.oshai.kotlinlogging.KotlinLogging
-import io.scheduler.comparison.quartz.jobs.state.ChunkedJobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.ChunkedJobMetadata
 import org.springframework.transaction.annotation.Transactional
 
 abstract class PaginatedJobHandlerBase<T, V : ChunkedJobMetadata, K> : PaginatedJobHandler<T, V, K> {
@@ -20,6 +20,7 @@ abstract class PaginatedJobHandlerBase<T, V : ChunkedJobMetadata, K> : Paginated
             return
         }
 
+        // Todo: fetch pages inside the same transaction.
         paginator.forEach { handleNextPage(it) }
         log.info { "[${orderJobMetadata.jobName}] Completed successfully" }
     }
