+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/handlers/streaming/impl/LocaLolaChunkedStreamJobHandler.kt
@@ -4,8 +4,8 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 import io.scheduler.comparison.quartz.domain.OrderRefund
 import io.scheduler.comparison.quartz.jobs.JobHandlerNames
 import io.scheduler.comparison.quartz.jobs.handlers.streaming.ChunkedStreamJobHandlerBase
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobMetadata
 import io.scheduler.comparison.quartz.messaging.LocaLolaRefundsSender
 import io.scheduler.comparison.quartz.repositories.streaming.LocaLolaStreamingFailuresRepository
 import org.springframework.context.annotation.Profile
