+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/pagination/JobPaginator.kt
@@ -1,6 +1,6 @@
 package io.scheduler.comparison.quartz.jobs.pagination
 
-import io.scheduler.comparison.quartz.jobs.state.ChunkedJobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.ChunkedJobMetadata
 
 interface JobPaginator<out T , out V : ChunkedJobMetadata, K> : Iterator<List<K>> {
