--- scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/JobMetadata.kt
@@ -1,7 +1,0 @@
-package io.scheduler.comparison.quartz.jobs.state
-
-interface JobMetadata {
-
-    val jobName: String
-
-}
