+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/JobState.kt
@@ -1,0 +1,10 @@
+package io.scheduler.comparison.quartz.jobs.state
+
+import io.scheduler.comparison.quartz.jobs.state.data.JobMetadata
+
+interface JobState<out T, out V : JobMetadata> {
+
+    val jobData: T
+    val jobMetadata: V
+
+}
