+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/data/ChunkedJobMetadata.kt
@@ -1,4 +1,4 @@
-package io.scheduler.comparison.quartz.jobs.state
+package io.scheduler.comparison.quartz.jobs.state.data
 
 interface ChunkedJobMetadata : JobMetadata {
