+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/data/JobMetadata.kt
@@ -1,0 +1,7 @@
+package io.scheduler.comparison.quartz.jobs.state.data
+
+interface JobMetadata {
+
+    val jobName: String
+
+}
