+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/repositories/pagination/LocaLolaFailuresRepository.kt
@@ -1,7 +1,7 @@
 package io.scheduler.comparison.quartz.repositories.pagination
 
 import io.scheduler.comparison.quartz.domain.OrderRefund
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobData
 import org.intellij.lang.annotations.Language
 import org.springframework.context.annotation.Profile
 import org.springframework.jdbc.core.simple.JdbcClient
