+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/repositories/streaming/LocaLolaStreamingFailuresRepository.kt
@@ -1,8 +1,8 @@
 package io.scheduler.comparison.quartz.repositories.streaming
 
 import io.scheduler.comparison.quartz.domain.OrderRefund
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobMetadata
 import io.scheduler.comparison.quartz.repositories.DomainRowMappers.orderRefundRowMapper
 import org.intellij.lang.annotations.Language
 import org.springframework.context.annotation.Profile
