+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/repositories/streaming/WildFruitStreamingOperationOnOrderRepository.kt
@@ -2,8 +2,8 @@ package io.scheduler.comparison.quartz.repositories.streaming
 
 import io.scheduler.comparison.quartz.domain.OperationOnOrder
 import io.scheduler.comparison.quartz.domain.OrderOperationStatus
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobData
-import io.scheduler.comparison.quartz.jobs.state.DedicatedOrderJobMetadata
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobData
+import io.scheduler.comparison.quartz.jobs.state.data.impl.DedicatedOrderJobMetadata
 import io.scheduler.comparison.quartz.repositories.DomainRowMappers.operationOnOrderRowMapper
 import org.intellij.lang.annotations.Language
 import org.springframework.context.annotation.Profile
